Fix C4018: cast GetMenuItemCount() to UINT in PageDisplay.cpp loop
menu_item is unsigned int (required by GetMenuString/GetMenuItemID) but
GetMenuItemCount() returns signed int, causing a signed/unsigned mismatch
warning.  Cast the count to UINT to match the loop variable type.

+++ src/PageDisplay.cpp
@@ -914,7 +914,7 @@ void CPageDisplay::SelectStatisticByName(int button, const char *statistic_name)
 			return;
 		}
 
-		for (menu_item = 0; menu_item < pSubPopup->GetMenuItemCount(); menu_item++) {
+		for (menu_item = 0; menu_item < (UINT)pSubPopup->GetMenuItemCount(); menu_item++) {
 			CString title;
 
 			pSubPopup->GetMenuString(menu_item, title, MF_BYPOSITION);
